Include import static statements in refactoring of package name
